+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v5.1
 - Better detection and naming of the Amazon, HP, PocketBook, Manta and Prestigio devices
 - Massive change of the OS Version for all Windows variants. Much cleaner now.
 - Detect the new Gecko fork Goanna (used by Palemoon)
+- Get the DeviceBrand from the URL/Email if no other is available
 
 v5.0
 ===